+++ test/Ed25519VerificationKey2020.spec.ts
@@ -6,10 +6,6 @@ import { base58btc } from '../src/baseX.js'
 import { mockKey, seed } from './mock-data.js'
 import * as multibase from 'multibase'
 import * as multicodec from 'multicodec'
-import { createRequire } from 'module'
-const require = createRequire(import.meta.url)
-const { Ed25519VerificationKey2018 } = require('@digitalbazaar/ed25519-verification-key-2018')
-
 import { Ed25519VerificationKey2020 } from '../src/index.js'
 
 // multibase base58-btc header
@@ -275,77 +271,6 @@ describe('Ed25519VerificationKey2020', () => {
       expect(result.verified).to.exist
       expect(result.verified).to.be.true
       expect(fingerprint).to.equal(fingerprint2)
-    })
-  })
-
-  describe('Backwards compat with Ed25519VerificationKey2018', () => {
-    const seedBytes = new TextEncoder().encode(seed).slice(0, 32)
-
-    it('2020 key should import from 2018', async () => {
-      const keyPair2018 = await Ed25519VerificationKey2018.generate({
-        seed: seedBytes,
-        controller: 'did:example:1234'
-      })
-
-      const keyPair2020 =
-        await Ed25519VerificationKey2020.fromEd25519VerificationKey2018({
-          keyPair: keyPair2018
-        })
-
-      // Both should have the same fingerprint
-      expect(keyPair2018.fingerprint()).to.equal(keyPair2020.fingerprint())
-
-      // Both should sign and verify the same
-      const data = new TextEncoder().encode('test data goes here')
-      const signatureBytes2018 = await keyPair2018.signer().sign({ data })
-
-      const signatureBytes2020 = await keyPair2020.signer().sign({ data })
-
-      expect(signatureBytes2018).to.eql(signatureBytes2020)
-      expect(
-        await keyPair2020
-          .verifier()
-          .verify({ data, signature: signatureBytes2018 })
-      ).to.be.true
-    })
-
-    it('2020 key should round trip serialize to 2018', async () => {
-      const keyPair2020 = await Ed25519VerificationKey2020.generate({
-        seed: seedBytes,
-        controller: 'did:example:1234'
-      })
-
-      const serialized2018 = keyPair2020.toEd255519VerificationKey2018({
-        publicKey: true,
-        privateKey: true
-      })
-
-      const parsedKeyPair = await Ed25519VerificationKey2020.from(
-        serialized2018
-      )
-
-      expect(keyPair2020.publicKeyMultibase).to.equal(
-        parsedKeyPair.publicKeyMultibase
-      )
-      expect(keyPair2020.publicKeyMultibase).to.equal(
-        parsedKeyPair.publicKeyMultibase
-      )
-    })
-
-    it('2020 key should generate the same from seed as 2018', async () => {
-      const keyPair2018 = await Ed25519VerificationKey2018.generate({
-        seed: seedBytes,
-        controller: 'did:example:1234'
-      })
-      const keyPair2020 = await Ed25519VerificationKey2020.generate({
-        seed: seedBytes,
-        controller: 'did:example:1234'
-      })
-
-      const data = new TextEncoder().encode('test data goes here')
-      const signatureBytes2018 = await keyPair2018.signer().sign({ data })
-      const signatureBytes2020 = await keyPair2020.signer().sign({ data })
-      expect(signatureBytes2018).to.eql(signatureBytes2020)
     })
   })